+++ coaster/sqlalchemy.py
@@ -3,7 +3,7 @@
 from __future__ import absolute_import
 from datetime import datetime
 import simplejson
-from sqlalchemy import Column, Integer, DateTime, Unicode, UnicodeText
+from sqlalchemy import Column, Integer, DateTime, Unicode, UnicodeText, CheckConstraint
 from sqlalchemy.sql import select, func
 from sqlalchemy.types import UserDefinedType, TypeDecorator, TEXT
 from sqlalchemy.orm import composite
@@ -114,16 +114,21 @@ class BaseNameMixin(BaseMixin):
     """
     #: Prevent use of these reserved names
     reserved_names = []
+    __name_blank_allowed__ = False
+    __name_length__ = __title_length__ = 250
 
     @declared_attr
     def name(cls):
         """The URL name of this object, unique across all instances of this model"""
-        return Column(Unicode(250), nullable=False, unique=True)
+        if cls.__name_blank_allowed__:
+            return Column(Unicode(cls.__name_length__), nullable=False, unique=True)
+        else:
+            return Column(Unicode(cls.__name_length__), CheckConstraint("name!=''"), nullable=False, unique=True)
 
     @declared_attr
     def title(cls):
         """The title of this object"""
-        return Column(Unicode(250), nullable=False)
+        return Column(Unicode(cls.__title_length__), nullable=False)
 
     def __init__(self, *args, **kw):
         super(BaseNameMixin, self).__init__(*args, **kw)
@@ -146,7 +151,7 @@ class BaseNameMixin(BaseMixin):
                 checkused = lambda c: bool(c in reserved or c in self.reserved_names or
                     self.__class__.query.filter_by(name=c).notempty())
             with self.__class__.query.session.no_autoflush:
-                self.name = unicode(make_name(self.title, maxlength=250, checkused=checkused))
+                self.name = unicode(make_name(self.title, maxlength=self.__name_length__, checkused=checkused))
 
 
 class BaseScopedNameMixin(BaseMixin):
@@ -165,16 +170,21 @@ class BaseScopedNameMixin(BaseMixin):
     """
     #: Prevent use of these reserved names
     reserved_names = []
+    __name_blank_allowed__ = False
+    __name_length__ = __title_length__ = 250
 
     @declared_attr
     def name(cls):
         """The URL name of this object, unique within a parent container"""
-        return Column(Unicode(250), nullable=False)
+        if cls.__name_blank_allowed__:
+            return Column(Unicode(cls.__name_length__), nullable=False)
+        else:
+            return Column(Unicode(cls.__name_length__), CheckConstraint("name!=''"), nullable=False)
 
     @declared_attr
     def title(cls):
         """The title of this object"""
-        return Column(Unicode(250), nullable=False)
+        return Column(Unicode(cls.__title_length__), nullable=False)
 
     def __init__(self, *args, **kw):
         super(BaseScopedNameMixin, self).__init__(*args, **kw)
@@ -196,7 +206,7 @@ class BaseScopedNameMixin(BaseMixin):
                 checkused = lambda c: bool(c in reserved or c in self.reserved_names or
                     self.__class__.query.filter_by(name=c, parent=self.parent).first())
             with self.__class__.query.session.no_autoflush:
-                self.name = unicode(make_name(self.short_title(), maxlength=250, checkused=checkused))
+                self.name = unicode(make_name(self.short_title(), maxlength=self.__name_length__, checkused=checkused))
 
     def short_title(self):
         """
@@ -223,15 +233,21 @@ class BaseIdNameMixin(BaseMixin):
     """
     Base mixin class for named objects with an id tag.
     """
+    __name_blank_allowed__ = False
+    __name_length__ = __title_length__ = 250
+
     @declared_attr
     def name(cls):
         """The URL name of this object, non-unique"""
-        return Column(Unicode(250), nullable=False)
+        if cls.__name_blank_allowed__:
+            return Column(Unicode(cls.__name_length__), nullable=False)
+        else:
+            return Column(Unicode(cls.__name_length__), CheckConstraint("name!=''"), nullable=False)
 
     @declared_attr
     def title(cls):
         """The title of this object"""
-        return Column(Unicode(250), nullable=False)
+        return Column(Unicode(cls.__title_length__), nullable=False)
 
     #: The attribute containing id numbers used in the URL in id-name syntax, for external reference
     url_id_attr = 'id'
@@ -244,7 +260,7 @@ class BaseIdNameMixin(BaseMixin):
     def make_name(self):
         """Autogenerates a :attr:`name` from the :attr:`title`"""
         if self.title:
-            self.name = unicode(make_name(self.title, maxlength=250))
+            self.name = unicode(make_name(self.title, maxlength=self.__name_length__))
 
     @property
     def url_id(self):
@@ -314,15 +330,21 @@ class BaseScopedIdNameMixin(BaseScopedIdMixin):
             parent = db.synonym('organizer')
             __table_args__ = (db.UniqueConstraint('organizer_id', 'url_id'),)
     """
+    __name_blank_allowed__ = False
+    __name_length__ = __title_length__ = 250
+
     @declared_attr
     def name(cls):
         """The URL name of this instance, non-unique"""
-        return Column(Unicode(250), nullable=False)
+        if cls.__name_blank_allowed__:
+            return Column(Unicode(cls.__name_length__), nullable=False)
+        else:
+            return Column(Unicode(cls.__name_length__), CheckConstraint("name!=''"), nullable=False)
 
     @declared_attr
     def title(cls):
         """The title of this instance"""
-        return Column(Unicode(250), nullable=False)
+        return Column(Unicode(cls.__title_length__), nullable=False)
 
     def __init__(self, *args, **kw):
         super(BaseScopedIdNameMixin, self).__init__(*args, **kw)
@@ -334,7 +356,7 @@ class BaseScopedIdNameMixin(BaseScopedIdMixin):
     def make_name(self):
         """Autogenerates a title from the name"""
         if self.title:
-            self.name = unicode(make_name(self.title, maxlength=250))
+            self.name = unicode(make_name(self.title, maxlength=self.__name_length__))
 
     @property
     def url_name(self):