+++ CHANGES.rst
@@ -4,6 +4,7 @@
 * Initial work on Fluentd logging.
 * New util: base_domain_matches compares if two domains have the same base domain.
 * utils.make_name now returns ASCII slugs instead of Unicode slugs.
+* New: domain_namespace_match function.
 
 0.4.2
 -----